Parser for reading data into the system
Ok so we will need to track customers and accounts and to do that we
will alsoneed to be able to import data from filesand then export it
back out to them. I have the user file pretty much figured out and
coded. the accounts class is a little more abscure at the moment. I can
read in data right now, but since some of the data is specific to
certain accounts I cannot do exactly what i did in the user file. what I
am thinking of is storing the data that is specific to all accounts in
ArrayLists, like I did in the users class, and then have an ArrayList of
objects to hold the account specific data